Add the solution of level-3 question-no.1
