+++ src/components/hexColorGenerator.jsx
@@ -1,14 +1,4 @@
-const hexaColor = () => {
-  let str = "0123456789abcdef";
-  let color = "";
-  for (let i = 0; i < 6; i++) {
-    let index = Math.floor(Math.random() * str.length);
-    color += str[index];
-  }
-  return "#" + color;
-};
-
-const HexaColor = () => {
+const HexaColor = ({hexaColor}) => {
   const hexaStyles = {
     lineHeight: "60px",
     textAlign: "center",
